+++ tools/tasks/seed/test.samples/serve.cover.samples.ts
@@ -1,17 +1,9 @@
-import * as browserSync from 'browser-sync';
-import * as chokidarSocketEmitter from 'chokidar-socket-emitter';
+import { createServer } from '../../../utils';
 import { join, sep } from 'path';
 import Config from '../../../config';
 
 export = () => {
 
-  chokidarSocketEmitter({
-    port: Config.HOT_LOADER_PORT,
-    path: join(Config.COVERAGE_DIR, Config.SAMPLES_DIR) + sep,
-    relativeTo: join(Config.COVERAGE_DIR, Config.SAMPLES_DIR) + sep,
-    dir: process.cwd()
-  });
-
-  browserSync.init(Config.BROWSER_SYNC_CONFIG_COVERAGE_SAMPLES);
+  createServer(join(Config.COVERAGE_DIR, Config.SAMPLES_DIR) + sep, Config.SERVER_SAMPLES_PORT)
 
 };